New implementation of .row() method
Also with a commend for testing: cmd/test_row_method.cpp
This version is now safe to use with any ImageType class, regardless of
whether it is accessed with direct IO, and Adapters.

All seems to check out OK. The following types of syntax are possible
(assuming images have size 3 along axis 3):

  out.row(3) = in.row(3);
  out.row(3) += in.row(3);
  out.row(3) -= in.row(3);

  Eigen::Matrix3f M;
  out.row(3) = M * Eigen::Vector3f(in.row(3));
  out.row(3) += M * Eigen::Vector3f(in.row(3)) + Eigen::Vector3f(in2.row(3));

  Eigen::VectorXd x (in.row(3));
  out.row(3) -= x;
  x -= in.row(3);

etc.

+++ cmd/dwidenoise.cpp
@@ -176,7 +176,7 @@ class DenoisingFunctor
       for (dwi.index(1) = pos[1]-extent[1]; dwi.index(1) <= pos[1]+extent[1]; ++dwi.index(1))
         for (dwi.index(0) = pos[0]-extent[0]; dwi.index(0) <= pos[0]+extent[0]; ++dwi.index(0), ++k)
           if (! is_out_of_bounds(dwi))
-            X.col(k) = dwi.row(3).template cast<float>();
+            X.col(k) = dwi.row(3);
     // reset image position
     dwi.index(0) = pos[0];
     dwi.index(1) = pos[1];
